lot of refactoring
