+++ api/service/delete_driver_test.go
@@ -7,12 +7,12 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	pb "github.com/ride-app/driver-service/api/gen/ride/driver/v1alpha1"
-	"github.com/ride-app/driver-service/mocks"
-	driverService "github.com/ride-app/driver-service/service"
+	driverService "github.com/ride-app/driver-service/api/service"
+	"github.com/ride-app/driver-service/testing/mocks"
 	"go.uber.org/mock/gomock"
 )
 
-var _ = Describe("GetVehicle", func() {
+var _ = Describe("DeleteDriver", func() {
 	var (
 		ctrl            *gomock.Controller
 		mockDriverRepo  *mocks.MockDriverRepository
@@ -32,19 +32,17 @@ var _ = Describe("GetVehicle", func() {
 	})
 
 	JustBeforeEach(func() {
-		mockVehicleRepo.EXPECT().GetVehicle(gomock.Any(), gomock.Any(), gomock.Any()).Return(&pb.Vehicle{}, nil)
+		mockDriverRepo.EXPECT().DeleteDriver(gomock.Any(), gomock.Any(), gomock.Any()).Return(nil, nil)
 	})
 
 	AfterEach(func() {
 		ctrl.Finish()
 	})
 
-	It("should get the vehicle successfully", func() {
-		req := connect.NewRequest(&pb.GetVehicleRequest{
-			Name: "drivers/valid-driver-id/vehicles/valid-vehicle-id",
-		})
+	It("should delete the driver successfully", func() {
+		req := connect.NewRequest(&pb.DeleteDriverRequest{Name: "drivers/valid-driver-id"})
 
-		_, err := service.GetVehicle(context.Background(), req)
+		_, err := service.DeleteDriver(context.Background(), req)
 		Expect(err).To(BeNil())
 	})
 })
