+++ api/service/get_status_test.go
@@ -7,12 +7,12 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	pb "github.com/ride-app/driver-service/api/gen/ride/driver/v1alpha1"
-	"github.com/ride-app/driver-service/mocks"
-	driverService "github.com/ride-app/driver-service/service"
+	driverService "github.com/ride-app/driver-service/api/service"
+	"github.com/ride-app/driver-service/testing/mocks"
 	"go.uber.org/mock/gomock"
 )
 
-var _ = Describe("UpdateVehicle", func() {
+var _ = Describe("GetStatus", func() {
 	var (
 		ctrl            *gomock.Controller
 		mockDriverRepo  *mocks.MockDriverRepository
@@ -32,24 +32,19 @@ var _ = Describe("UpdateVehicle", func() {
 	})
 
 	JustBeforeEach(func() {
-		mockVehicleRepo.EXPECT().UpdateVehicle(gomock.Any(), gomock.Any(), gomock.Any()).Return(&pb.Vehicle{}, nil)
+		mockDriverRepo.EXPECT().GetStatus(gomock.Any(), gomock.Any(), gomock.Any()).Return(&pb.Status{}, nil)
 	})
 
 	AfterEach(func() {
 		ctrl.Finish()
 	})
 
-	It("should update the vehicle successfully", func() {
-		req := connect.NewRequest(&pb.UpdateVehicleRequest{
-			Vehicle: &pb.Vehicle{
-				Name:         "drivers/valid-driver-id/vehicles/valid-vehicle-id",
-				Type:         pb.Vehicle_TYPE_ERICKSHAW,
-				DisplayName:  "Erickshaw",
-				LicensePlate: "KA-01-1234",
-			},
+	It("should get the status successfully", func() {
+		req := connect.NewRequest(&pb.GetStatusRequest{
+			Name: "drivers/valid-driver-id",
 		})
 
-		_, err := service.UpdateVehicle(context.Background(), req)
+		_, err := service.GetStatus(context.Background(), req)
 		Expect(err).To(BeNil())
 	})
 })
