+++ testing/mocks/vehicle_repository.go
